+++ tests/test_particle_contract_audit.sh
@@ -1226,6 +1226,12 @@ check_present_re '^run_mpi\(\)[[:space:]]*\{' scripts/docker-build-test.sh
 check_present 'timeout "\$TEST_TIMEOUT" "\$MPIRUN"' scripts/docker-build-test.sh
 check_present 'MPIRUN_FLAGS' scripts/docker-build-test.sh
 check_present 'MPIRUN_FLAGS: "--oversubscribe"' .github/workflows/ci.yml
+check_absent '^int (gridMask|logGrid);' \
+  tests/test_oh4_runtime_globals.c \
+  tests/test_oh4p_runtime_smoke.c \
+  tests/test_oh4p_capacity_guard.c \
+  tests/test_oh4s_runtime_smoke.c \
+  tests/test_oh4s_capacity_guard.c
 check_present_re '^expect_mpi_failure_log_contains\(\)[[:space:]]*\{' \
   scripts/docker-build-test.sh
 check_present_re '^expect_mpi_failure_log_cases\(\)[[:space:]]*\{' \